Auto-Inventory-Project/pyproject.toml
[project]
name = "auto-inventory-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=0.1.0",
    "chromadb>=1.0.15",
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.27.0",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langmem>=0.0.28",
    "langmem-adapter>=0.0.2",
    "openai-agents>=0.2.3",
    "psycopg2>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "pypdf>=5.8.0",
    "sqlalchemy>=2.0.41",
    "stripe>=12.3.0",
    "twilio>=9.7.0",
]
